huge changes 2.8.0 finalized
