+++ html/about.html
@@ -22,7 +22,6 @@
         body {
             font-family: 'Lato', sans-serif;
         }
-
         @keyframes color-rotate{
             from{
                 filter: hue-rotate(0deg);
@@ -32,11 +31,11 @@
             }
         }
 
-        /* img{
+        #logo{
             animation: color-rotate 1s;
             animation-iteration-count: infinite;
             animation-direction: alternate;
-        } */
+        }
 
 
 
@@ -53,31 +52,24 @@
 
 <body>
     <div style="text-align: center;">
-        <img src="../figs/charts.png" width="75" height="75">
+        <img id='logo' src="../figs/charts.png" width="75" height="75">
         <div style="font-weight: bold;font-size: xx-large;">Interactive Data Editor</div>
         <div style="font-size: large;font-style: italic;margin-top: 3px; font-weight:500">
-            A Software to interactively edit data in a graphical manner.
+            A Software to interactively edit data in a graphical manner
         </div>
+        <div id='versionname' style="font-size: large;margin-top: 7px">Version 2.8.0 </div>
         <div style="font-size: larger;margin-top: 5px;font-style: italic;margin-bottom: 9px; font-weight:600">
             Developed by Koushik Naskar
         </div>
-        <div id='versionname' style="font-size: large;margin-top: 7px">Version 2.7.0</div>
-        <div style="font-size: larger;margin-top: 2pc;font-style: italic;margin-bottom: 9px">
-            Copyright &copy; November, 2019 Koushik Naskar
-        </div>
-        <div style="font-size: larger; margin-bottom: 10px; font-weight: 600; font-style:italic">
-            <a href="javascript:void(0)" onclick="op('https://koushikphy.github.io/');">koushikphy.github.io</a>
+        <div style="font-size: larger;margin-top: 50px;font-style: italic;margin-bottom: 9px;font-weight:530">
+            Copyright &copy; April, 2020 Koushik Naskar
         </div>
 
-
-
-
-
-        <a href="javascript:void(0)" onclick="op('https://www.facebook.com/koushik.naskar3');"><img title="Reach me on Facebook"               src="../figs/facebook.png" width="30" height="30"></a>
-        <a href="javascript:void(0)" onclick="op('https://github.com/Koushikphy');"><img title="Follow me on Github" src="../figs/github.png"                width="30" height="30"></a>
-        <a href="javascript:void(0)"            onclick="op('https://mail.google.com/mail/?view=cm&fs=1&to=koushik.naskar9@gamil.com&su=[Interactive Data Editor]');"><img  title="Email Me"              src="../figs/gmail.png" width="30" height="30"></a>
-
-
+        <a title="Open website" href="javascript:void(0)" onclick="shell.openExternal('https://koushikphy.github.io');"><img src="../figs/internet.png" width="30" height="30"></a>
+        <a title="Follow me Facebook" href="javascript:void(0)" onclick="shell.openExternal('https://www.facebook.com/koushik.naskar3');"><img src="../figs/facebook.png" width="30" height="30"></a>
+        <a title="Follow me on Github" href="javascript:void(0)" onclick="shell.openExternal('https://github.com/Koushikphy');"><img src="../figs/github.png"            width="30" height="30"></a>
+        <a title="Reach me on Gamil" href="javascript:void(0)" onclick="shell.openExternal('https://mail.google.com/mail/?view=cm&fs=1&to=koushik.naskar9@gamil.com&su=[Interactive Data Editor]');"><img src="../figs/gmail.png" width="30" height="30"></a>
+        
 
         <div style="font-size: larger;font-weight:bold;margin-top: 3.8pc;margin-bottom: 5px">Powered By:</div>
         <a href="javascript:void(0)"  onclick="op('https://nodejs.org/');"><img title="NodeJS" src="../figs/node.png" width="100" height="60"></a> <br>
