+++ html/data.html
@@ -19,6 +19,8 @@
 &nbsp;3.<a href="javascript:void(0)" onclick="save('../Data/Morse.dat');">Morse Function</a><br>
 &nbsp;4.<a href="javascript:void(0)" onclick="save('../Data/3D2.dat');">Sin(sqrt(x**2+y**2))/sqrt(x**2+y**2)</a><br>
 &nbsp;5.<a href="javascript:void(0)" onclick="save('../Data/3D2.dat');">Sin(x)*Cos(y)</a><br>
+&nbsp;6.<a href="javascript:void(0)" onclick="save('../Data/MultiColumn2D.dat');">2D Multi Column</a><br>
+&nbsp;7.<a href="javascript:void(0)" onclick="save('../Data/MultiColumn3D.dat');">3D Multi Column</a><br>
 <script >
 	const fs = require("fs");        
 	const { remote } = require('electron');
