--- html/settings.html
@@ -1,163 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>Interactive Data Editor - Documentation</title>
-    <!-- <link rel="stylesheet" href="../lib/pure-min.css"> -->
-    <!-- <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bulma@0.8.0/css/bulma.min.css"> -->
-    <!-- <link rel="stylesheet" href="../../font-awesome-4.7.0/css/font-awesome.css"> -->
-    <style>
-
-    .row{
-        /* margin-bottom: 5px; */
-        border-style: groove;
-        display:inline-block;
-        width: 75%;
-        padding: 5px;
-        text-align: left;
-    }
-    .index{
-        font-family: Cambria, Cochin, Georgia, Times, 'Times New Roman', serif;
-        font-size: larger;
-        margin-right: 5px;
-    }
-    .filename{
-        font-family:'Times New Roman', Times, serif;
-        font-size: larger;
-        width: 65%;
-        display: inline-block;
-        overflow: auto;
-        white-space: nowrap;
-    }
-
-    button:focus{
-        outline: none;
-    }
-
-    .tools{
-        float: right;
-        display: inline;
-    }
-    .colselect{
-        display: inline;
-    }
-    .radio{
-        margin-right: 2px;
-        margin-left: 5px;
-    }
-    .lButton{
-        width: 4pc;
-        font-size: large;
-        margin: 5px;
-    }
-
-
-
-
-
-    .copyBtn,.clsBtn{
-    border: none;
-    background: none;
-    text-align: right;
-    font: right;
-    padding-left: 2px;
-    padding-right: 2px;
-    height: 1.5pc;
-    width: 1.5pc;
-}
-
-.clsBtn:hover{
-    background-color: rgb(233, 98, 98);
-}
-.copyBtn:hover{
-    background-color: rgb(179, 223, 252);
-}
-
-    #full-container{
-        border: 2px groove #aaa;
-        border-radius: 1px;
-    }
-    #close{
-        float: right;
-    }
-
-
-    </style>
-    <script>
-        if (typeof module === "object") {
-            window.module = module;
-            module = undefined;
-        }
-    </script>
-</head>
-
-<body>
-    <div id = 'full-container'>
-        <button id='close'>close this</button>
-        <br>
-        <div id='main-container' style="text-align: center;">
-    </div>
-    <script src="../lib/jquery.min.js"></script>
-    <script>
-        var fileList = [
-            'File 1 egwef',
-            "File 2",
-            "File 3"
-        ]
-        var fileCols = [
-            [0,1,2],
-            [2,3,4],
-            [1,2,3]
-        ]
-        var checked = 0
-
-
-        function buildDOM(){
-            document.getElementById('main-container').innerHTML = 
-                fileList.map((i,v)=>createCard(i,v)).join('')+
-                `<button class='lButton' title=''>Ok</button>
-                <button class='lButton' title=''>Close</button>
-                `
-        document.getElementsByClassName('radio')[checked].checked = true
-        }
-        buildDOM()
-
-
-        function tools(option,index){
-            console.log(option,index)
-            if(option==0){ //select editable
-                checked = index
-            }else if (option==1) { // clone this
-                fileList.push( fileList[index])
-            }else if(option==2) { // close this
-                fileList.splice(index,1)
-                if(index<checked) checked--
-            }
-            buildDOM()
-        }
-
-        function createCard(fileName, index){
-            let shortName = fileName
-            return `<div class="row">
-                <label class="index">${index+1}.</label>
-                <div class="tools">
-                    <input  class='radio' type="radio" onclick="tools(0,${index})" title='Select this as editable'>
-                    <button class="copyBtn" onclick="tools(1,${index})" title='Use this file'>
-                        <svg viewBox="0 0 1792 1792">
-                            <path d="M1664 1632v-1088q0-13-9.5-22.5t-22.5-9.5h-1088q-13 0-22.5 9.5t-9.5 22.5v1088q0 13 9.5 22.5t22.5 9.5h1088q13 0 22.5-9.5t9.5-22.5zm128-1088v1088q0 66-47 113t-113 47h-1088q-66 0-113-47t-47-113v-1088q0-66 47-113t113-47h1088q66 0 113 47t47 113zm-384-384v160h-128v-160q0-13-9.5-22.5t-22.5-9.5h-1088q-13 0-22.5 9.5t-9.5 22.5v1088q0 13 9.5 22.5t22.5 9.5h160v128h-160q-66 0-113-47t-47-113v-1088q0-66 47-113t113-47h1088q66 0 113 47t47 113z"/>
-                        </svg>
-                        </button>
-                    <button class="clsBtn" onclick="tools(2,${index})" title='Remove this file'>
-                        <svg viewBox="0 0 1792 1792">
-                            <path d="M1490 1322q0 40-28 68l-136 136q-28 28-68 28t-68-28l-294-294-294 294q-28 28-68 28t-68-28l-136-136q-28-28-28-68t28-68l294-294-294-294q-28-28-28-68t28-68l136-136q28-28 68-28t68 28l294 294 294-294q28-28 68-28t68 28l136 136q28 28 28 68t-28 68l-294 294 294 294q28 28 28 68z"/>
-                        </svg>
-                    </button>
-                    </div>
-                    <label class="filename" title='${fileName}'>${shortName}</label>
-            </div><br>`
-        }
-
-    </script>
-</body>
-</html>
