--- todos
@@ -1,6 +1,0 @@
-cursor when drag and when not
-align of the popup boxes
-tickangle of ticks in angle
-
-
-points onclick add the css to points and $('.js-plotly-plot .plotly .cursor-crosshair')
